reorganize features: move from functonality to features
- Rename src/functonality/ to src/features/ (fix typo)
- Moved Web3Provider, countdownTimer, and other core features
- Better semantic organization for feature modules
- Consistent naming convention
