+++ packages/dapp/src/features/README.md
@@ -1,0 +1,17 @@
+# Features Directory
+
+## Overview
+
+Feature-level components and helpers used across the dapp (web3, ticket purchase, countdown, and email templates).
+
+## Contents
+
+- `Web3Provider.tsx`
+- `customTheme.tsx`
+- `ticketpurchasecomponent.tsx`
+- `eventEscrowComponent.tsx`
+- `countdownTimer.ts`
+- `walletStatus.tsx`
+- `navBarController.tsx`
+- `mockfunding.ts` (dev helper)
+- `emailNewsletter/main.html`
