+++ packages/dapp/src/features/ticketpurchasecomponent.tsx
@@ -1,8 +1,8 @@
 // TicketPurchaseComponent.tsx
 import React, { useEffect, useState } from 'react';
 import { useSession } from 'next-auth/react';
-import { TicketPurchaseProps } from '@/utils/dev/typeInit';
-import { CONTRACT_ADDRESSES, contracts } from '@/utils/prod/contractInit';
+import { TicketPurchaseProps } from '@/types/contracts';
+import { CONTRACT_ADDRESSES, contracts } from '@/utils/contractInit';
 import useExhibit from '@/lib/useGetExhibitById';
 import {
    useWeb3Provider,
